+++ public/admin/_app/immutable/nodes/24.1egJQ4q8.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/CZ4CZXnO.js";export{e as component};